requirements.txt
django
django-cors-headers
channels
channels-redis
redis
python-dotenv
loguru
django-prometheus
uvicorn
gunicorn
kombu
faker
psycopg
psycopg2-binary
celery
djangorestframework
adrf
asgiref
websockets
wsproto
django-celery-results